Add static Webfinger file generation
When running a single-author site, it's relatively common to render a
static Webfinger file for the single account on the site. Using static
mode will enable this behavior.

It's not technically correct via the RFC for two reasons:

1. It does not account for checking other subjects, even pages on the
   same site. This means that it will allow you to respond correctly
   for, say, Mastodon, but not correctly if someone wants to check the
   copyright status of a page on your site.
2. It does not filter the links by `rel` in the JSON Resource Descriptor
   since it renders all of them to the file.

These caveats can be acceptable depending on what you want to use
Webfinger for.