Update expected Ruby and Bundler versions
Ruby 2.7 will be out of even security patch range [next month][1] so
there's not much point in supporting it. Ruby 3.0 will go to
security-only maintenance at the same time, but will likely have another
year of security support.

Bundler 2 has been out for three years so bumping to that with the bump
to Ruby 3.0 seems ideal.

[1]: https://bugs.ruby-lang.org/projects/ruby/wiki/ReleaseEngineering